implement for & while loops
three things:

1) i deviated from the book here; it has you implement a while loop, and
then implement the for loop by desugaring it to a while loop.  i ended
up doing the reverse: implementing the for loop, and then implementing a
while loop as a for loop with no initializer or increment.  this will be
useful later when implementing the break & continue loop control
statements.

2) this is the first statement in the interpreter where i have to reuse
expressions and statements (everywhere else they're one-shot).  this
means i have to clone the condition, increment, and body, which is lame.
i'll need to refactor the entire interpreter to pass around references
so we don't "consume" expressions/statements/values when using them.

3) in order to avoid leaking the initialier's variable declaration into
the enclosing scope, i wrap the for statement in a block statement.